+++ sdmx-dl-provider-util/src/test/java/internal/http/curl/CurlTest.java
@@ -12,6 +12,8 @@ import java.io.InputStreamReader;
 import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.nio.charset.StandardCharsets;
+import java.util.List;
+import java.util.TreeMap;
 
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -36,23 +38,24 @@ public class CurlTest {
         try (InputStream stream = Resource.getResourceAsStream(CurlTest.class, "curlhead.txt").orElseThrow(IOException::new)) {
             try (BufferedReader reader = new BufferedReader(new InputStreamReader(stream, StandardCharsets.UTF_8))) {
                 assertThat(Curl.CurlHead.parseResponse(reader))
-                        .isEqualTo(Curl.CurlHead
-                                .builder()
-                                .code(200)
-                                .message("OK")
-                                .header("Date", singletonList("Wed, 20 Oct 2021 10:58:37 GMT"))
-                                .header("Expires", singletonList("-1"))
-                                .header("Cache-Control", singletonList("private, max-age=0"))
-                                .header("Content-Type", singletonList("text/html; charset=ISO-8859-1"))
-                                .header("P3P", singletonList("CP=\"This is not a P3P policy! See g.co/p3phelp for more info.\""))
-                                .header("Server", singletonList("gws"))
-                                .header("X-XSS-Protection", singletonList("0"))
-                                .header("X-Frame-Options", singletonList("SAMEORIGIN"))
-                                .header("Accept-Ranges", singletonList("none"))
-                                .header("Vary", singletonList("Accept-Encoding"))
-                                .header("Transfer-Encoding", singletonList("chunked"))
-                                .build()
-                        );
+                        .isEqualTo(new Curl.CurlHead(
+                                new Curl.Status(200, "OK"),
+                                new TreeMap<String, List<String>>(String.CASE_INSENSITIVE_ORDER) {
+                                    {
+                                        put("Date", singletonList("Wed, 20 Oct 2021 10:58:37 GMT"));
+                                        put("Expires", singletonList("-1"));
+                                        put("Cache-Control", singletonList("private, max-age=0"));
+                                        put("Content-Type", singletonList("text/html; charset=ISO-8859-1"));
+                                        put("P3P", singletonList("CP=\"This is not a P3P policy! See g.co/p3phelp for more info.\""));
+                                        put("Server", singletonList("gws"));
+                                        put("X-XSS-Protection", singletonList("0"));
+                                        put("X-Frame-Options", singletonList("SAMEORIGIN"));
+                                        put("Accept-Ranges", singletonList("none"));
+                                        put("Vary", singletonList("Accept-Encoding"));
+                                        put("Transfer-Encoding", singletonList("chunked"));
+                                    }
+                                }
+                        ));
             }
         }
     }